requirements.txt
asgiref==3.6.0
Django==4.1.6
django-ckeditor==6.5.1
django-crispy-forms==2.0
django-extensions==3.2.1
django-js-asset==2.0.0
django-tinymce==3.5.0
Pillow==9.4.0
pydotplus==2.0.2
pyparsing==3.0.9
sqlparse==0.4.3
tzdata==2022.7
